moved initialise_gauss_profilers to profilers
